Generate an app based on a local Phoenix framework checkout
To generate an app based on a local checkout of Phoenix, use, use the
hidden `--dev` flag while running the `phx.new` generator. The command
needs to run from the `installer` directory inside a the local checkout,
because the resulting project will link to the local checkout using
relative paths.

  $ git clone https://github.com/phoenixframework/phoenix.git
  $ cd phoenix/installer

Use the `--dev` flag while running the installer. If you have Phoenix
installed on your machine, the installer will default to using that.
This causes an error, as the used installer’s path doesn’t match the
current one, which is required when using the `--dev` flag.

  $ mix phx.new fishtank --no-ecto --no-gettext --umbrella --dev
  ** (Mix) --dev projects must be generated inside Phoenix directory

Remove the installed version of Phoenix to force the installer to use
the generator from the local checkout, resolving the installer-path
mismatch.

  $ mix archive | grep phx_new
  * phx_new-1.4.12
  $ mix archive.uninstall phx_new-1.4.12
  Are you sure you want to uninstall /Users/jeff/.asdf/installs/elixir/master/.mix/archives/phx_new-1.4.12? [Yn]
  y
  $ mix phx.new fishtank --no-ecto --no-gettext --umbrella --dev
  Compiling 7 files (.ex)
  * creating fishtank_umbrella/.gitignore
  * creating fishtank_umbrella/config/config.exs
  * creating fishtank_umbrella/config/dev.exs
  ...

This creates a Phoenix app based on the latest version of Phoenix. The
generated project depends on the local Phoenix checkout by using a
relative path in the `mix.exs` file.

``` elixir
defmodule FishtankWeb.MixProject do
  use Mix.Project

  # ...

  defp deps do
    [
      {:phoenix, path: "../../../..", override: true},
      {:phoenix_html, "~> 2.11"},
      {:phoenix_live_reload, "~> 1.2", only: :dev},
      {:telemetry_metrics, "~> 0.4"},
      {:telemetry_poller, "~> 0.4"},
      {:fishtank, in_umbrella: true},
      {:jason, "~> 1.0"},
      {:plug_cowboy, "~> 2.0"}
    ]
  end

  # ...
end
```