Generate a Phoenix edge application
Generating a Phoenix application using the latest version from Github
is done by first creating an application based on a local install of the
Phoenix framework, and then switching to using the Github version.

## Generate an app based on a local Phoenix framework checkout

To generate an app based on a local checkout of Phoenix, use, use the
hidden `--dev` flag while running the `phx.new` generator. The command
needs to run from the `installer` directory inside a the local checkout,
because the resulting project will link to the local checkout using
relative paths.

  $ git clone https://github.com/phoenixframework/phoenix.git
  $ cd phoenix/installer

Use the `--dev` flag while running the installer. If you have Phoenix
installed on your machine, the installer will default to using that.
This causes an error, as the used installer’s path doesn’t match the
current one, which is required when using the `--dev` flag.

  $ mix phx.new fishtank --no-ecto --no-gettext --umbrella --dev
  ** (Mix) --dev projects must be generated inside Phoenix directory

Remove the installed version of Phoenix to force the installer to use
the generator from the local checkout, resolving the installer-path
mismatch.

  $ mix archive | grep phx_new
  * phx_new-1.4.12
  $ mix archive.uninstall phx_new-1.4.12
  Are you sure you want to uninstall /Users/jeff/.asdf/installs/elixir/master/.mix/archives/phx_new-1.4.12? [Yn]
  y
  $ mix phx.new fishtank --no-ecto --no-gettext --umbrella --dev
  Compiling 7 files (.ex)
  * creating fishtank_umbrella/.gitignore
  * creating fishtank_umbrella/config/config.exs
  * creating fishtank_umbrella/config/dev.exs
  ...

This creates a Phoenix app based on the latest version of Phoenix. The
generated project depends on the local Phoenix checkout by using a
relative path in the `mix.exs` file.

``` elixir
defmodule FishtankWeb.MixProject do
  use Mix.Project

  # ...

  defp deps do
    [
      {:phoenix, path: "../../../..", override: true},
      {:phoenix_html, "~> 2.11"},
      {:phoenix_live_reload, "~> 1.2", only: :dev},
      {:telemetry_metrics, "~> 0.4"},
      {:telemetry_poller, "~> 0.4"},
      {:fishtank, in_umbrella: true},
      {:jason, "~> 1.0"},
      {:plug_cowboy, "~> 2.0"}
    ]
  end

  # ...
end
```

## Use Phoenix edge from Github

Instead of relying on a local checkout of Phoenix being available,
switch the phoenix dependency in `mix.exs` to allow the project to
download the framework directly from Github by updating the Phoenix
dependency line.

``` elixir
# apps/fishtank_web/mix.exs
defmodule FishtankWeb.MixProject do
  use Mix.Project

  # ...

  defp deps do
    [
      {:phoenix, github: "phoenixframework/phoenix", override: true},
      {:phoenix_html, "~> 2.11"},
      {:phoenix_live_reload, "~> 1.2", only: :dev},
      {:telemetry_metrics, "~> 0.4"},
      {:telemetry_poller, "~> 0.4"},
      {:fishtank, in_umbrella: true},
      {:jason, "~> 1.0"},
      {:plug_cowboy, "~> 2.0"}
    ]
  end

  # ...
end

## Non-local npm dependency paths

Getting the Phoenix dependency directly from Github instead of a
relative path to a local directory allows moving the project out of the
`installer` directory in the Phoenix source.

    $ mv ~/opensource/phoenix/installer/fishtank_umbrella ~/opensource/fishtank_umbrella
    $ cd ~/opensource/fishtank_umbrella
    $ mix phx.server
    [info] Running FishtankWeb.Endpoint with cowboy 2.7.0 at 0.0.0.0:4000 (http)
    [error] Could not start node watcher because script "/Users/jeff/opensource/fishtank_umbrella/apps/fishtank_web/assets/node_modules/webpack/bin/webpack.js" does not exist. Your Phoenix application is still running, however assets won't be compiled. You may fix this by running "cd assets && npm install".
    [info] Access FishtankWeb.Endpoint at http://localhost:4000
    ...

To get assets to compile automatically when running the app, the npm
dependencies need to be installed by running `npm install` in the
project’s assets directory.

    $ cd apps/fishtank_web/assets
    $ npm install
    npm ERR! code ENOLOCAL
    npm ERR! Could not install from "../../../../.." as it does not contain a package.json file.
    ...

The install produces an error, as the `phoenix` and `phoenix_html`
dependencies in the `package.json` file still point to relative paths in
the local Phoenix checkout (`"file:../../../../../"` and
`"file:../../../deps/phoenix_html"`, respectively)

Instead, use paths to the `deps` directory, which holds the Phoenix
framework version now downloaded from GitHub.

``` js
apps/fishtank_web/assets/package.json
{
  "repository": {},
  "description": " ",
  "license": "MIT",
  "scripts": {
    // ...
  },
  "dependencies": {
    "phoenix": "file:../../../deps/phoenix",
    "phoenix_html": "file:../../../deps/phoenix_html"
  },
  "devDependencies": {
    // ...
  }
}
```

NOTE: This example uses an umbrella project. Use `../deps/phoenix` and
`../deps/phoenix_html` for non-umbrella applications.

With the updated paths, npm now installs the dependencies from the
dependencies downloaded to the `deps` directory instead of trying to
find the local Phoenix checkout.

    $ cd apps/fishtank_web/assets
    $ npm install
    ...
    npm notice created a lockfile as package-lock.json. You should commit this file.
    added 778 packages from 329 contributors and audited 8112 packages in 23.965s
    found 0 vulnerabilities