Set up SpectatorChannel
The generated `SpectatorChannel` includes stubs to handle multiple
rooms, pings and authorization. Since the spectator channel is in a
single room that's publicly available, this code is unneeded.

A minimal implementation of the `SpectatorChannel` only has a function
that handles joins.

``` elixir
defmodule FishtankWeb.SpectatorChannel do
  use FishtankWeb, :channel

  def join("fishtank", _message, socket) do
    {:ok, socket}
  end
end
```

The `SpectatorChannelTest` has a function that checks if joining the
channel returns a `Phoenix.Socket`:

``` elixir
defmodule FishtankWeb.SpectatorChannelTest do
  setup do
    {:ok, _, socket} =
      FishtankWeb.UserSocket
      |> socket()
      |> subscribe_and_join(FishtankWeb.SpectatorChannel, "fishtank")

    %{socket: socket}
  end

  test "broadcasts are pushed to the client", %{socket: socket} do
    broadcast_from!(socket, "broadcast", %{"some" => "data"})
    assert_push "broadcast", %{"some" => "data"}
  end

  test "joins the channel", %{socket: socket} do
    assert %Phoenix.Socket{
             joined: true,
             channel: FishtankWeb.SpectatorChannel,
             handler: FishtankWeb.UserSocket
           } = socket
  end
end
```